requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
aws-cdk-lib==2.178.1
aws-cdk.asset-awscli-v1==2.2.223
aws-cdk.asset-kubectl-v20==2.1.4
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==39.2.20
boto3==1.36.19
botocore==1.36.19
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
cloudflare==4.0.0
constructs==10.4.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
jmespath==1.0.1
jsii==1.106.0
publication==0.0.3
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.3.0
